requirements.txt
pre-commit==3.7.1
pandas==2.2.2
langchain==0.2.1
langchain_community==0.2.1
python-dotenv==1.0.1
openai==1.30.5
chromadb==0.5.0
tiktoken==0.7.0
